+++ src/admin/fetchAdmin/DisplayOrders.tsx
@@ -1,5 +1,4 @@
 import React, { useState } from "react";
-import Popthis from "./Popthis"; // Assuming Popthis is in the parent directory
 import { Link } from "react-router-dom";
 import Head from "../../components/Head";
 import Button from "../../components/Button";
@@ -24,6 +23,11 @@ interface Order {
 
 interface DisplayOrdersProps {
   orders: Order[];
+}
+interface PopupProps {
+  order: Order;
+  onClose: () => void;
+ 
 }
 
 const DisplayOrders: React.FC<DisplayOrdersProps> = ({ orders }) => {
@@ -54,6 +58,44 @@ const DisplayOrders: React.FC<DisplayOrdersProps> = ({ orders }) => {
     setShowPopthis(false);
     setSelectedOrder(null);
   };
+  
+ const Popthis: React.FC<PopupProps> = ({ order,onClose }) => {
+  console.log(order);
+  return (   <>
+  <div className="fixed inset-0 flex items-center justify-center z-50">
+        {/* Overlay */}
+        <div className="absolute inset-0 bg-mywhite bg-opacity-50"></div>
+        {/* Popup */}
+        <div className="relative p-8 rounded-lg shadow-lg w-1/2 bg-mywhite text-black">
+          <button
+            onClick={onClose}
+            className="absolute top-0 right-0 mt-4 mr-4 text-xl"
+          >
+            &times;
+          </button>
+        {order.product.map((product) => ( 
+      <>  
+          <h2 className="font-bold mb-4 text-2xl flex items-center justify-center text-myred">Products</h2><div className="mb-4 text-lg">
+                   <strong>desc</strong> {product.desc}
+               </div><div className="mb-4 text-lg">
+                       <strong> Name:</strong> {product.name}
+                   </div><div className="mb-4 text-lg">
+                       <strong>Price:</strong> ${product.price}
+                   </div><div className="mb-4 text-lg">
+                       <strong>Quantity:</strong> {product.price}
+                   </div><div className="mb-4 text-lg">
+                       <strong>Rating:</strong> {product.quantity}
+                   </div>
+      </>
+  ))} <div className="mb-4 text-lg">
+  <strong>Status:</strong> {order.status}
+</div>
+</div>
+</div>
+  </>
+    )
+  }
+
 
   return (
     <>
@@ -125,6 +167,7 @@ const DisplayOrders: React.FC<DisplayOrdersProps> = ({ orders }) => {
       )}
     </>
   );
+  
 };
 
 export default DisplayOrders;
